# docs/openapi.yaml
openapi: '3.0.3'
info:
  title: Warehouse Management System - OpenAPI 3.0
  version: 0.1.1
  description: WIP
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost/api/v1
...
paths:
...
  /warehouse/{token}:
    get:
      tags:
        - warehouses
      summary: Get warehouse information
      description: Get information about warehouse
      operationId: getWarehouse
      security:
        - cookieAuth: [ ]
      parameters:
        - name: token
          in: path
          description: "Warehouse ID"
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Warehouse information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WarehouseResponse'
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenResponse'
        "404":
          description: Warehouse not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                notFound:
                  value:
                    code: 404
                    type: error
                    message: "Warehouse not found"
    put:
      tags:
        - warehouses
      summary: Update warehouse information
      description: Update warehouse information
      operationId: updateWarehouse
      security:
        - cookieAuth: [ ]
      parameters:
        - name: token
          in: path
          description: "Warehouse ID"
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Update warehouse
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WarehouseCreationRequest'
      responses:
        "200":
          description: Updated warehouse
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WarehouseResponse'
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenResponse'
        "404":
          description: Warehouse not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                notFound:
                  value:
                    code: 404
                    type: error
                    message: "Warehouse not found"
    delete:
      tags:
        - warehouses
      summary: Delete warehouse
      description: Delete warehouse
      operationId: deleteWarehouse
      security:
        - cookieAuth: [ ]
      parameters:
        - name: token
          in: path
          description: "Warehouse ID"
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Warehouse deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                success:
                  value:
                    code: 200
                    type: success
                    message: "Warehouse deleted"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenResponse'
        "404":
          description: Warehouse not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                notFound:
                  value:
                    code: 404
                    type: error
                    message: "Warehouse not found"


  # Stock
  /stock/warehouse/{warehouse_id}:
    get:
      tags:
        - stock
      summary: Get all stock
      description: WIP
      operationId: listStock
      security:
        - cookieAuth: [ ]
      parameters:
        - name: warehouse_id
          in: path
          description: "Warehouse ID"
          required: true
          schema:
            type: integer
            format: int64
        - name: page
          in: query
          description: "Page number"
          required: false
          schema:
            type: integer
            format: int32
            example: 1
        - name: limit
          in: query
          description: "Items per page"
          required: false
          schema:
            type: integer
            format: int32
            example: 10
        - name: sort
          in: query
          description: "Sort by"
          required: false
          schema:
            type: string
            example: "name"
        - name: search
          in: query
          description: "Search by name"
          required: false
          schema:
            type: string
            example: "Apple"
      responses:
        "200":
          description: List of stock
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StockResponse'
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenResponse'
        "404":
          description: Warehouse not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                notFound:
                  value:
                    code: 404
                    type: error
                    message: "Warehouse not found"
    post:
      tags:
        - stock
      summary: Add new stock
      description: WIP
      operationId: addStock
      security:
        - cookieAuth: [ ]
      parameters:
        - name: warehouse_id
          in: path
          description: "Warehouse ID"
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Add stock
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StockRequest'
      responses:
        "201":
          description: Created stock
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockResponse'
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenResponse'
        "404":
          description: Warehouse not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                notFound:
                  value:
                    code: 404
                    type: error
                    message: "Warehouse not found"
  /stock/warehouse/{warehouse_id}/stock/{stock_id}:
    get:
      tags:
        - stock
      summary: Get stock information
      description: Get information about stock
      operationId: getStock
      security:
        - cookieAuth: [ ]
      parameters:
        - name: warehouse_id
          in: path
          description: "Warehouse ID"
          required: true
          schema:
            type: integer
            format: int64
        - name: stock_id
          in: path
          description: "Stock ID"
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Stock information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockResponse'
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenResponse'
        "404":
          description: Stock not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                notFound:
                  value:
                    code: 404
                    type: error
                    message: "Stock not found"
    put:
      tags:
        - stock
      summary: Update stock information
      description: Update stock information
      operationId: updateStock
      security:
        - cookieAuth: [ ]
      parameters:
        - name: warehouse_id
          in: path
          description: "Warehouse ID"
          required: true
          schema:
            type: integer
            format: int64
        - name: stock_id
          in: path
          description: "Stock ID"
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Update stock
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StockRequest'
      responses:
        "200":
          description: Updated stock
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockResponse'
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenResponse'
    delete:
      tags:
        - stock
      summary: Delete stock
      description: Delete stock
      operationId: deleteStock
      security:
        - cookieAuth: [ ]
      parameters:
        - name: warehouse_id
          in: path
          description: "Warehouse ID"
          required: true
          schema:
            type: integer
            format: int64
        - name: stock_id
          in: path
          description: "Stock ID"
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Stock deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                success:
                  value:
                    code: 200
                    type: success
                    message: "Stock deleted"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenResponse'
        "404":
          description: Stock not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                notFound:
                  value:
                    code: 404
                    type: error
                    message: "Stock not found"

  # Notifications
...
components:
  schemas:
...
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        type:
          type: string
        message:
          type: string
      xml:
        name: "##default"
    TimeInfo:
      type: object
      properties:
        created_at:
          type: integer
          format: int64
          example: 1735120168
        updated_at:
          type: integer
          format: int64
          example: 1735120168
    Location:
      type: object
      properties:
        latitude:
          type: number
          format: double
          example: 51.090709
        longitude:
          type: number
          format: double
          example: 71.418143
    WarehouseResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
          example: 12
        name:
          type: string
          example: Central
        location:
          $ref: '#/components/schemas/Location'
        time_info:
          $ref: '#/components/schemas/TimeInfo'
...
    WarehouseCreationRequest:
      type: object
      properties:
        name:
          type: string
          example: Main
        location:
          $ref: '#/components/schemas/Location'
    StockResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
          example: 12
        name:
          type: string
          example: "Apple"
        description:
          type: string
          example: "Fresh apples"
        quantity:
          type: integer
          format: int32
          example: 100
        unit_price:
          type: number
          format: double
          example: 1.99
        unit_ammount:
          type: number
          example: 100
        total_price:
          type: number
          format: double
          example: 199.0
        warehouse_id:
          type: integer
          format: int32
          example: 12
        properties:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
                example: "Color"
              value:
                type: string
                example: "Red"
        time_info:
          $ref: '#/components/schemas/TimeInfo'
    StockRequest:
      type: object
      properties:
        name:
          type: string
          example: "Apple"
        description:
          type: string
          example: "Fresh apples"
        quantity:
          type: integer
          format: int32
          example: 100
        unit_price:
          type: number
          format: double
          example: 1.99
        unit_ammount:
          type: number
          example: 100
        warehouse_id:
          type: integer
          format: int32
          example: 12
        properties:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
                example: "Color"
              value:
                type: string
                example: "Red"
...
    UnauthorizedResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
          example: 401
        type:
          type: string
          example: error
        message:
          type: string
          example: "You are not authenticated"
      xml:
        name: "##default"
    ForbiddenResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
          example: 403
        type:
          type: string
          example: error
        message:
          type: string
          example: "You don't have permission to access this resource"
      xml:
        name: "##default"
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth
security:
  - cookieAuth: [ ]